requirements.txt
asgiref==3.2.10
attrs==20.2.0
colorama==0.4.4
Django==3.1.2
django-cors-headers==3.5.0
django-seed==0.2.2
djangorestframework==3.12.1
docopt==0.6.2
Faker==4.14.0
importlib-metadata==2.0.0
iniconfig==1.1.1
packaging==20.4
pathtools==0.1.2
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
pytest-watch==4.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.0
text-unidecode==1.3
toml==0.10.1
watchdog==0.10.3
zipp==3.4.0
